% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{age.process}
\alias{age.process}
\title{Processing age to print its distribution}
\usage{
age.process(age)
}
\arguments{
\item{age}{a vector of age, each entry is either a number (like 34) or age by decade (like 30s)}
}
\value{
each age is either repeated 10 times or expanded to 10 numbers (for example, 30s is expanded to 30, 31, ..., 39).
}
\description{
Processing age to print its distribution
}
