% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood_inference.R
\name{bets.likelihood.conditional}
\alias{bets.likelihood.conditional}
\title{(Profile) Conditional likelihood given B and E}
\usage{
bets.likelihood.conditional(
  params,
  data,
  M = Inf,
  r = NULL,
  params_init = NULL
)
}
\arguments{
\item{params}{A vector of parameters (names: r, ip_q50, ip_q95)}

\item{data}{A data frame with three columns: B, E, S}

\item{M}{Right truncation for symptom onset}

\item{r}{Parameter for epidemic growth (overrides \code{{params}})}

\item{params_init}{Initial parameters for computing the profile likelihood}
}
\value{
Conditional log-likelihood.
}
\description{
(Profile) Conditional likelihood given B and E
}
\keyword{internal}
