% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BETS.R
\docType{package}
\name{BETS}
\alias{BETS}
\title{BETS: A package for analyzing early epidemic data}
\description{
The BETS package provides likelihood inference for early epidemic data with four key epidemiological events: Beginning of exposure, End of exposure, time of Transmission, and time of Symptom onset. It jointly estimates the epidemic doubling time and incubation period and is able to correct for different kinds of sample selection.
}
\references{
Qingyuan Zhao, Nianqiao Ju, and Sergio Bacallado. "BETS: The dangers of selection bias in early analyses of the coronavirus disease (COVID-19) pandemic", 2020.
}
