% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{date.process}
\alias{date.process}
\title{Transform date to numeric}
\usage{
date.process(date)
}
\arguments{
\item{date}{a vector of dates of the form "DD-MMM" (for example, 23-Jan).}
}
\value{
a vector of days since December 1st, 2019 (or example, 23-Jan is converted to 23+31 = 52).
}
\description{
Transform date to numeric
}
