% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{multiple.impute.onset}
\alias{multiple.impute.onset}
\title{Multiple imputation of symptom onset date}
\usage{
multiple.impute.onset(data, m = 50)
}
\arguments{
\item{data}{a data frame with the following columns: Arrived, Symptom, Initial, Hospital, Confirmed.}
}
\value{
a matrix of symptom onset dates with missing values imputed using the mice package.
}
\description{
Multiple imputation of symptom onset date
}
