% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_infected.R
\name{parse.one.infected}
\alias{parse.one.infected}
\title{Parse infected date (basic)}
\usage{
parse.one.infected(infected)
}
\arguments{
\item{infected}{A string of the form "DATE1" or "DATE1 to DATE2".}
}
\value{
A vector of length 2 for the infection window
}
\description{
Parse infected date (basic)
}
\keyword{internal}
