% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrape_data.R
\name{scrape.shenzhen}
\alias{scrape.shenzhen}
\title{Partially generate Shenzhen data}
\usage{
scrape.shenzhen(file)
}
\description{
Partially generate Shenzhen data
}
\examples{

data <- scrape.shenzhen("../data-raw/shenzhen_raw_data.txt")
head(data)

}
