% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate.case}
\alias{simulate.case}
\title{Simulate case information from the generative BETS model}
\usage{
\method{simulate}{case}(
  n = 1e+07,
  params = c(pi = 0.1, lambda_w = 5e-04, lambda_v = 0.001, kappa = 4e-08, r = 0.2, alpha
    = 9, beta = 1.5, L = 53.5)
)
}
\description{
Simulate case information from the generative BETS model
}
\keyword{internal}
