% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wuhan_exported}
\alias{wuhan_exported}
\title{COVID-19 exported from Wuhan}
\format{A data frame with 378 rows and 4 variables:
\describe{
  \item{Location}{Where the case is confirmed.}
  \item{Gender}{Gender of the patient.}
  \item{Age}{Age of the patient.}
  \item{B}{Beginning of stay in Wuhan.}
  \item{E}{End of stay in Wuhan.}
  \item{S}{Symptom onset.}
}}
\usage{
wuhan_exported
}
\description{
Constructed from \code{covid19_data}, see \code{example(preprocess.data)}.
}
\keyword{datasets}
