% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood_inference.R
\name{bets.likelihood.unconditional}
\alias{bets.likelihood.unconditional}
\title{Approximate profile likelihood}
\usage{
bets.likelihood.unconditional(params, data, params_init = NULL, L = NULL)
}
\arguments{
\item{params}{A vector of parameters (names: rho, r, ip_q50, ip_q95)}

\item{data}{A data frame with three columns: B, E, S}

\item{params_init}{Initial parameters for computing the profile likelihood}

\item{L}{Day of travel quarantine}
}
\value{
When \code{params} contains all the parameters (rho, r, ip_q50, ip_q95), returns the approximate log-likelihood of \code{data}. When \code{params} contains some but not all the parameters, returns the profile log-likelihood.
}
\description{
Approximate profile likelihood
}
\keyword{internal}
