% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_infected.R
\name{parse.infected}
\alias{parse.infected}
\title{Parse the infected date}
\usage{
parse.infected(data)
}
\arguments{
\item{data}{a data frame with the following columns: Infected, Arrived, Symptom, Initial, Confirmed.}
}
\value{
the data frame with two new columns, Infected_first and Infected_last
}
\description{
Deprecated. Used in the previous analysis, now replaced by \code{preprocess.data}.
}
\keyword{internal}
