% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_infected.R
\name{preprocess.data}
\alias{preprocess.data}
\title{Prepare data frame for analysis}
\usage{
preprocess.data(data, infected_in = c("Wuhan", "Outside"))
}
\arguments{
\item{data}{A data frame}

\item{infected.in}{Either "Wuhan" or "Outside"}
}
\value{
A data frame
}
\description{
Prepare data frame for analysis
}
\details{
here is a summary of the procedures
\enumerate{
  \item Convert all dates to number of days since 1-Dec-2019.
  \item Restrict to cases with a known symtom onset date.
  \item Separates data into those returned from Wuhan and those infected outside of wuhan.
  \item Parse column 'Infected' into two columns: Infected_first and Infected_last.
  \item For all cases, set Infected_first to 1 if it is missing.
  \item For outside cases, set Infected_last to be no later than symptom onset.
  \item For Wuhan-exported cases, set Infected_last to no later than symptom onset and arrival.
}
}
\examples{

data <- cases.in.china
head(preprocess.data(data))

}
\author{
Nianqiao Ju <nju@g.harvard.edu>
}
